seed delete if tables exist

+++ server/seed.py
@@ -1,5 +1,6 @@
 from app import app
 from models import db, User, Project, Task, Part, Tool, TaskTools
+from sqlalchemy import inspect
 
 import random as r
 
@@ -9,12 +10,20 @@ fake = Faker()
 
 with app.app_context():
 
-    User.query.delete()
-    Project.query.delete()
-    Task.query.delete()
-    Part.query.delete()
-    Tool.query.delete()
-    TaskTools.query.delete()
+    existing_tables = set(inspect(db.engine).get_table_names())
+
+    if "user" in existing_tables:
+        User.query.delete()
+    if "project" in existing_tables:
+        Project.query.delete()
+    if "task" in existing_tables:
+        Task.query.delete()
+    if "part" in existing_tables:
+        Part.query.delete()
+    if "tool" in existing_tables:
+        Tool.query.delete()
+    if "task_tools" in existing_tables:
+        TaskTools.query.delete()
 
     demo_user = User(username="demo")
     demo_user.password_hash = demo_user.username + "pass"
